Fix "wrong-type-argument" and case sensitivity
In some buffers where there is no file, such as `*scratch*`, "wrong-type-argument
stringp nil" is fired back and prevents the buffer from being killed or closing
Emacs. This is because `buffer-file-name` returns nil, and `file-name-extension`
does not like it thereby throwing a wrench in the works.

Additionally, some files may have different capitalization for their extensions,
such as ".R" (uppercase is recommend by some) for R language. Because of this,
the previous assoc statement would erroneously return nil instead of `(r ."R")`,
and continue on to the backup case where it uses the mode name for the language.
Switching to `assoc-string` and enabling case folding eliminates the issue
allowing the return of the expected language no matter the casing of the
extension string.